requirements.txt
Flask>=0.8
gevent-socketio>=0.2.3